net: fec: prevent enet counter from being stopped on link status changes
The Ethernet MAC was being entirely reset on a link down event, which
resulted in the ENET counter being stopped, and real-time stopping on the
system as well.
Fix this by only stopping MAC Rx/Tx operation (ETHEREN==0) as is already
done on the SoloX, instead of doing a full MAC reset.

Signed-off-by: Yannick Vignon <yannick.vignon@nxp.com>
Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>