net: fec: apply workaround for ERR007885 for normal path too
IMX6SX ENET module has a potential sequencing issue when running
with multiple transmit queues (ERR007885) which causes a transmit
queue to hang. A workaround is to delay the write of the TDAR
register and is already applied to the TSO path.

Apply the same workaround for the normal transmit path to prevent
the TX queues from blocking.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>