net: mscc: ocelot: make ocelot_mact_read return DST_IDX
We'll need to export this function to Felix DSA, who will use it
specifically for the purpose of knowing the destination PGID
corresponding to a MAC table entry.

Unfortunately, the way this function works now is that the PGID is
specified as an input parameter, then the function reads from hardware
the DST_IDX associated with the entry, then compares the input parameter
with what was read from hardware, and gives an error on mismatch.

Stop these guessing games and just let the only caller thus far,
ocelot_fdb_dump(), deal with that.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>