net: mscc: ocelot: add port force forwarding setting
The ocelot port will remove from forwarding mask when it's down. But the
FRER function need to keep the ocelot port in forwarding mask to keep
its splitting function working.

This patch add an ocelot_bridge_force_forward_port() function to set the
port in force forwarding mode, and export the function for DSA driver.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>