net: dsa: felix: change frer stream port to be forced forwarding
Original codes use a FRER PGID to set the FRER stream to be forwarded to
all split ports, which restricts all split ports to have the same VLAN
forwarding mask as the source port, or all split ports must be in the
forwarding mask.

This patch change the FRER stream forwarding port to be forced forward,
So that it is still in the forwarding mask after the port is down.

This is a workaround to fix the hardware issue: If the frame cannot be
forwarded, the FRER splitting module will not work.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>