net: stmmac: fix kernel panic when removing vlan_prio rule
When removing a tc-flower entry, the flow_cls_offload struct passed to the
driver is only supposed to be used to retrieve the flow cookie, as the
other fields are not initialized. Since tc_del_vlan_flow() was relying on
those fields, this was triggering a kernel panic when removing a vlan_prio
entry.

Maintain a table of active vlan_prio entries in the driver so the
information to remove them is available when needed.

Fixes: 8e46ec21360fcbd40c9f3b913c40683aa4dafb13("net: stmmac: add RX frame steering based on VLAN priority in tc flower")

Signed-off-by: Yannick Vignon <yannick.vignon@nxp.com>