net: dsa: felix: update basetime of tas when ptp adjust
When ptp clock adjust, the basetime of the TAS configuration will become
unreliable. We need reset the TAS configuration by using a new basetime.

For example, if the driver get a basetime 0 of Qbv configuration from
user, and current time is 20000. The driver will set the TAS basetime to
be 20000. After PTP clock adjust, the current time becomes to be 10000.
If the TAS basetime is still 20000, it will be a future time, and TAS
entry list will stop running. Another example, if the current time
becomes to be 10000000 after PTP clock adjust, a large time offset can
cause the hardware to hang.

This patch introduce a tas_clock_adjust() function to reset the TAS
module by using a new basetime after the PTP clock adjust. This can
avoid issues above.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>