arm: dts: add avb device tree for imx6ull-14x14-evk board
This dts will add essentially the Media Clock Recovery (Internal Method) for the board.
- GPT2 will be used for the AVB HW timer and MCR mechanism while GPT1 will be kept, as by default, for the clock system.

For this board, the eth0 is mapped to the ENET2 (eth1 to ENET1) : that's why
we connected the ENET2 To the GPT while passing ethernet port 0 (eth0)  as rec-channel
in the gpt node.

To make it work, a HW rework is needed (described in the dts).

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>