arm: dts: imx6ull: avb: set peripheral clock setting to IPG instead of 24MHz oscillator
The peripheral clock is set (in imx6ul_clocks_init()) to the external crystal (24MHz) to
add low power idle support. But, this change brings many limitations throughout the system.
Also to make the MCR mechanism work properly, the IMX6ULL Reference Manual (30.2.1) states
that the GPT_CLK (The external clock operating the GPT counter and set to 12.288 MHz
for our MCR mechanism) needs to be less than 1/4 of frequency of the peripheral clock.

Let the clock to its original parent and frequency (66 MHz).

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>