net: ocelot: clean tx_skbs when set HWTSTAMP_TX_ON
timestamp ID will not match to the skb stored in tx_skbs when it's
blocked and dropped by Qbv gate. Because timestamp ID only has a 0-3
value in sillicon, this will cause a wrong match between timestamp ID
and skb frame which is wait to transmitted. Clean the tx_skbs each time
when HWTSTAMP_TX_ON is set can recover this timestamp ID not match issue.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>