net: dsa: sja1105: use mqprio mapping to configure port egress pcp mapping
For egress queue selection, the sja1105 has two mappings: "Ingress to 
Egress PCP mapping" per port (first entries in the L2 forwarding table) 
and "Egress PCP to Hardware queue" mapping per pcp value.
Currently, the driver is hardcoding a 1:1 mapping for both of them and 
there is no direct way to configure that mapping from userspace for 
different usecases: like applying the priority to traffic class mapping 
from IEEE Std 802.1Q-2018 for AVB usecases.

For that, implement the TC_SETUP_QDISC_MQPRIO offload in the driver and 
apply the configured mapping as Egress PCP to Queue mapping (while 
keeping the 1:1 mapping for the Ingress to Egress). On qdisc deletion, 
the default 1:1 mapping is restored. On the other hand, since the 
mapping (from both tc-mqprio and tc-taprio) is now offloaded to 
hardware, we don't need to check for the skb queue mapping in 
sja1105_xmit() and we can pass on the skb priority as pcp value and let 
the hardware select the proper egress queue. Furthermore, the PCP in the 
DSA header has only 3 bits thus skb priority, processed as pcp value, 
will have its upper bits truncated for values higher that 7 but still 
covers the pcp values, ranging from 0 to 7, we're using.

Signed-off-by: Alexandre Howard <alexandre.howard@nxp.com>