dsa: netc: add mqprio set support
TC taprio or mqprio needs to map the priority to traffic class. This
patch add the support to map the priority to traffic class in hardware.

When switch port in standalone mode, the priority of skb will use the
map in Linux network stack. There is no need to offload the map in
hardware. When the switch port is in bridge mode, the map will be load
into hardware.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>