LF-14857-3 net: enetc: add jumbo frames support for ENETC v4
Confirmed with NETC IP team, with the default memory on i.MX95, there is
no issue with supporting jumbo frames functionally. But if the use-case
requires jumbo frames and small size frames on all ports, running at
line-rates, there might be performance issues. But this is corner case,
we should not drop support for jumbo frames because of this, and some
customers have jumbo frames requirement in DPDK. So add the support for
jumbo frames to ENETC v4.

Signed-off-by: Wei Fang <wei.fang@nxp.com>