net: netc: support free-running counter for rx and tx timestamps
The free-running timer being now used by default for packets timestamping
the 64bits timestamps are constructed by reading the FRT hardware registers.

All timestamps are expressed based on the free-running counter. So if the
skb has not the USE_CYCLES flags, meaning it is not binded to a virtual clock,
then the socket is expecting timestamp in the synchronized time domain, hence
netc_ptp_timer_convert is called.

Signed-off-by: Laurent Brando <laurent.brando@nxp.com>