ptp: extended support for timestamp conversion IOCTL
Adding support for virtual clock to physical clock conversion
when the physical clock is not free-running.

When getcycles64() function is implemented by the PHC driver
the physical clock the virtual clocks are created from remains
adjustable. A new callback converttime() is added to the
ptp_clock_info structure allowing the underlaying PHC driver
to perform ptp time conversion between the virtual clock domain
and the synchronized clock domain.

Signed-off-by: Laurent Brando <laurent.brando@nxp.com>
Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>