arm64: dts: imx93-14x14-sja1105-avb: remove reset gpio property for fec
The device tree has recently changed by removing a reset gpio from the eqos
node and moving it to earlier boot stage causing a change in the network interface
probe order: making eth0 as the eqos and eth1 as the fec.

This is a workaround to keep the network interface probe order in Linux kernel
into the same order and eliminate the deferred probe in fec caused by the phy reset:
it basically relies on u-boot to do the phy reset (which is done by default)
and remove the need to do that in the Linux driver. It should be removed
later once the avb endpoint stack starts supporting dynamic port mapping in
the kernel module

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>