net: dsa: felix: delay port hang workaround if taprio schedule is pending
vsc9959_tas_guard_bands_update() has multiple purposes, one of them is
to update QSYS_QMAXSDU_CFG per port-tc to conform to the queueMaxSDU
specified by the user, or in lack of that, to prevent port hangs. The
latter motivation comes from a non-conformant hardware implementation,
where frames that are larger than the minimum gate interval for their
traffic class are never eligible for TX scheduling. Normally, IEEE
802.1Q recommends that such frames are aged out on TX by the bridge
port, according to clause 6.5.6 Frame lifetime. However this mechanism
is broken with this hardware IP, because the frames are not eligible for
TX scheduling in the first place.

The distinction between these 2 reasons of updating QSYS_QMAXSDU_CFG is
important, because the queueMaxSDU is a user-visible property, whereas
the port hang workaround should be invisible to the user.

Specifically, queueMaxSDU is defined by IEEE 802.1Q as a managed object
for scheduled traffic which is not contingent on the existence of a
currently operational schedule (otherwise said, this managed object is
not replicated into an "admin" and an "oper" variant, like others are).
When a queueMaxSDU value is configured, it is enforced right away.

On the other hand, the driver workaround for port hangs should take place
only once a schedule is active, because it is a transparent schedule fixup
in itself. But this is not done, and vsc9959_tas_guard_bands_update()
applies QSYS_QMAXSDU_CFG right away, irrespective of the origin of the
programmed values.

Due to this fact, it can be noticed that when applying a schedule with
very short intervals and a queueMaxSDU array of zeroes (i.e. no
restrictions), the hardware port drops packets right away even if the
schedule has a future base time, and thus is not yet operational. We have
been let to know that this behavior is not correct.

Therefore, modifications to the driver logic are necessary in the
following areas:
 - vsc9959_tas_guard_bands_update() must refrain from calling
   vsc9959_port_qmaxsdu_set() with values derived from the port schedule,
   if the port schedule itself is not active. However, it must still write
   values coming from the user.
 - A delayed work item must be introduced, which checks whether there is a
   pending schedule or not (there seems to be no IRQ emitted for this
   event). For robustness and ease of implementation, this work item must
   reschedule itself every second, because the PTP clock can suffer time
   jumps in between the moment the taprio admin schedule is installed, and
   the moment it is supposed to become oper (i.e. its base time). Thus, a
   naive procedure of calculating a single delay value, of the base time
   minus the current PTP time, would not tolerate clock jumps, or would
   need complex logic to readjust during clock jumps. With the "check
   QSYS_PARAM_STATUS_REG_8_CONFIG_PENDING once per second" method, we can
   guarantee that the delayed work item will reapply the QSYS_QMAXSDU_CFG
   values no later than one second after the base time.

There will be a software-based delay between the moment the admin schedule
takes effect and the updating of QSYS_QMAXSDU_CFG no matter what we do.
During the interval of inconsistency, frames which can never be sent will
accumulate in the switch queues and buffers. The hope is that the transient
nature of this issue does not negatively impact users.

Co-developed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>