Revert "net: fec: keep ptp timer running regardless of link status"
With the commit a1477dc87d ("net: fec: Restart PPS after link state change"), this change
is incomplete and causes an issue on platforms that does a full MAC reset
in fec_ctrl_reset(). Revert it to prepare for a proper fix, under CONFIG_AVB_SUPPORT to
guarantee that the ptp clock is never stopped and we will not need any save/restore
ptp state.

This reverts commit 928d109907d35afb7eefb417ae1ac869f273880b.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>